+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fix comment sorting
 - Issue where images could not be viewed
+- Error when post has no comments
 
 ### Changes